requirements.txt
langchain
google-generativeai
langchain-google-genai
python-dotenv
pillow
langchain_openai
langchain_experimental
langchain_community
langchain_huggingface
langgraph
pymysql 
chromadb
sentence-transformers
pandas
streamlit
pydantic
langchain_groq
flask
cryptography==3.4.8